backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.11",
    "openai-whisper>=20240930",
    "python-multipart>=0.0.20",
    "setuptools-rust>=1.10.2",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.9",
    "websockets>=15.0",
]

[tool.pytest.ini_options]
pythonpath = "."
